[step2] refactor: LottoProcessor 클래스 삭제 및 Lotto, LottoWinningStats 생성자에서 각 인스턴스변수 초기화하도록 수정
