+++ src/main/java/lotto/views/InputView.java
@@ -2,7 +2,7 @@ package lotto.views;
 
 import java.util.Scanner;
 
-import lotto.LottoProcessor;
+import lotto.LottoWinningStats;
 
 public class InputView {
     private static final Scanner SCANNER = new Scanner(System.in);
@@ -11,7 +11,7 @@ public class InputView {
         System.out.println("구입 금액을 입력해주세요.");
         int purchaseAmount = SCANNER.nextInt();
         SCANNER.nextLine();
-        System.out.println((purchaseAmount / LottoProcessor.PRICE) + "개를 구매했습니다.");
+        System.out.println((purchaseAmount / LottoWinningStats.PRICE) + "개를 구매했습니다.");
         return purchaseAmount;
     }
