+++ src/test/java/lotto/numbers/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.numbers;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -9,15 +9,11 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import lotto.numbers.AllNumbers;
-import lotto.numbers.Lotto;
-
 public class LottoTest {
     @ParameterizedTest
     @ValueSource(ints = {1, 3, 5, 10, 20})
     public void generateNumbers(int purchaseCount) {
         Lotto lotto = new Lotto(purchaseCount);
-        lotto.generateNumbers(purchaseCount);
 
         assertThat(lotto.getNumbersSize()).isEqualTo(purchaseCount);
         lotto.getNumbers().forEach(numbers -> assertThat(numbers.size()).isEqualTo(Lotto.TOTAL_NUMBERS_SIZE));
@@ -27,7 +23,6 @@ public class LottoTest {
     @MethodSource("containsNumberParam")
     public void containsNumber(int index, int number) {
         Lotto lotto = new Lotto(10);
-        lotto.generateNumbers(10);
         assertThat(lotto.containsNumber(index, number)).isInstanceOfAny(Boolean.class);
     }
 
@@ -35,7 +30,6 @@ public class LottoTest {
     @ValueSource(ints = {1, 3, 5, 10, 20})
     public void getNumbers(int purchaseCount) {
         Lotto lotto = new Lotto(purchaseCount);
-        lotto.generateNumbers(purchaseCount);
 
         assertThat(lotto.getNumbers()).isNotEmpty();
         assertThat(AllNumbers.getNumbers().containsAll(lotto.getNumbers().get(0))).isEqualTo(true);
